HDFS-8043. NPE in MiniDFSCluster teardown. Contributed by Brahma Reddy Battula.
(cherry picked from commit 5459b241c86cc9a26fecca9a06ceaf524e48fed4)

+++ hadoop-hdfs-project/hadoop-hdfs/src/test/java/org/apache/hadoop/hdfs/MiniDFSCluster.java
@@ -1748,11 +1748,14 @@ public class MiniDFSCluster {
         nameNode = null;
       }
     }
-    if (deleteDfsDir) {
+    if (base_dir != null) {
+      if (deleteDfsDir) {
         base_dir.delete();
-    } else {
+      } else {
         base_dir.deleteOnExit();
+      }
     }
+
   }
   
   /**
